backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

# --- AI / STT ---
# openai-whisper==20231117       # install separately: pip install openai-whisper
# torch==2.2.0                   # install separately: pip install torch

pyannote.audio==3.1.1
soundfile==0.12.1
librosa==0.10.1

# --- NLP ---
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
spacy==3.7.4

# --- LLM Providers ---
openai>=1.0.0
groq>=0.4.2

# --- Database ---
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
alembic>=1.13.0

# --- Audio Processing ---
noisereduce>=3.0.0
pydub>=0.25.1

# --- Core ---
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.3
aiofiles==23.2.1
websockets>=12.0

# --- Testing (dev) ---
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# httpx>=0.27.0
# pytest-cov>=4.1.0
